installed WebIOPi
installed and configured WebIOPi, it will run on boot and HTTP server
will be available on the port 9009. - password protected of course,
maybe a read-only view of it can be put as a self-refreshing public page
(simple HTML)

still waiting for the connector to connect FAN and stuff to the
Expansion board, will test with Fan using some hacked connector soon if
delay is long.